requirements.txt
decorator==5.1.1
langchain==0.3.9
langchain_openai==0.2.10
openai==1.55.3
python-dotenv==1.0.1
tenacity==9.0.0
